Add JWT based authentication with RBAC

+++ app/src/dto/response/auth_dto/create_user_response.rs
@@ -6,7 +6,7 @@ use time::OffsetDateTime;
 #[builder(setter(into))]
 #[serde(rename_all = "camelCase")]
 pub struct CreateUserResponse {
-    id: i64,
+    id: i32,
 
     username: String,
